Popravi prevajalne napake iz CI: podrsanje, jezik, ura, smer premika
- SwipeToShift: horizontalDrag ima en parameter (sprememba prihaja iz
  positionChange().x), stanje potrebuje getValue/setValue za delegata.
- AgendaWidget in TimelineComponents: še zadnji mesti, ki sta klicali slovensko
  jezikovno konstanto namesto skupne izbire jezika.
- DisplayFormat: ura je del vzorcev po jeziku, ne literal mimo njih.
- TimelineAction.Shift sprejme minute kot Long, zato toLong() na podrsanju.

+++ app/src/main/java/com/example/mydailyroutine/app/presentation/RoutineApp.kt
@@ -310,7 +310,7 @@ fun RoutineApp(viewModel: RoutineViewModel, access: NotificationAccess,
                         // belongs to the month grid's own gestures.
                         modifier = Modifier.fillMaxSize().swipeToShift(
                             enabled = !state.panels.showGoals && data.mode != TimelineMode.YEAR,
-                            onShift = { direction -> onAction(TimelineAction.Shift(direction)) },
+                            onShift = { direction -> onAction(TimelineAction.Shift(direction.toLong())) },
                         ), transitionSpec = {
                             // Moving through time slides along a shared horizontal axis in the
                             // direction of travel; changing scale (day -> week) shares no geometry
